requirements.txt
numpy==1.22.2
opencv_contrib_python==4.5.4.58
Pillow==9.0.1
streamlit==1.5.0
streamlit_player==0.1.5
